+++ src/app/components/items/item-list/item-list.component.scss
