+++ src/app/components/items/item/item.component.scss
