--- src/app/home/home.module.ts
@@ -1,15 +1,0 @@
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {HomeComponent} from './components/home/home.component';
-import { ItemsModule } from '../items/items.module';
-
-@NgModule({
-  declarations: [HomeComponent],
-  exports: [HomeComponent],
-  imports: [
-    CommonModule,
-    ItemsModule
-  ]
-})
-export class HomeModule {
-}
