+++ src/app/services/items/items.service.ts
@@ -1,9 +1,9 @@
-import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
-import { Item } from 'src/entities/item.model';
-import { FirebaseService } from '../../services/firebase/firebase.service';
-import { CollectionType } from '../../services/firebase/models/collection-type.model';
-import { IFirebaseService } from '../../services/firebase/models/firebase-service.interface';
+import {Injectable} from '@angular/core';
+import {Observable} from 'rxjs';
+import {Item} from 'src/entities/item.model';
+import {FirebaseService} from '../firebase/firebase.service';
+import {CollectionType} from '../firebase/models/collection-type.model';
+import {IFirebaseService} from '../firebase/models/firebase-service.interface';
 
 @Injectable()
 export class ItemsService implements IFirebaseService<Item>{
